requirements.txt
# 核心Web框架
fastapi
uvicorn[standard]
python-multipart

# LangChain生态系统
langchain
langchain-community
langchain-core
langgraph

# AI模型和API
openai

# 数据处理和机器学习
numpy
faiss-cpu
Pillow

# 文档处理
unstructured
markdown
pypdf2
python-docx
openpyxl
xlrd

# 配置和环境
python-dotenv
pydantic

# 搜索工具
tavily-python

# 工具包
requests
aiofiles 